requirements.txt
asgiref==3.2.10
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
Django==3.0.8
django-autoslug==1.9.8
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.3.0
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.6.0
more-itertools==8.2.0
packaging==20.3
Pillow==8.1.0
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.8.1
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytz==2020.1
reportlab==3.5.59
requests==2.23.0
six==1.14.0
sorl-thumbnail==12.7.0
sqlparse==0.3.1
urllib3==1.25.9
UserManager==0.5.3
wcwidth==0.1.9
zipp==3.1.0
